Move formatDate to its own component
